Completely revamped loading system. Still needs to be heavily tested and adjusted.
